+++ R/check.derivatives.R
@@ -5,7 +5,7 @@
 # Author: Jelmer Ypma
 # Date:   24 July 2011
 #
-# Compare analytic derivatives wih finite difference approximations.
+# Compare analytic derivatives with finite difference approximations.
 #
 # Input:
 #    .x : compare at this point
@@ -27,8 +27,11 @@
 #   2013-10-27: Added relative_error and flag_derivative_warning to output list.
 #   2014-05-05: Replaced cat by message, so messages can now be suppressed by
 #         suppressMessages.
-#   2023-02-09: Cleanup and tweaks for safety and efficiency (AA)
-
+#   2023-02-09: Cleanup and tweaks for safety and efficiency (Avraham Adler).
+#   2024-07-03: Converted to use nl.grad to have package be consistent in its
+#         numerical differentiation routines and allow user choice between three
+#         and five point versions of the algorithm (Avraham Adler).
+#
 
 #' Check analytic gradients of a function using finite difference
 #' approximations
@@ -36,28 +39,32 @@
 #' This function compares the analytic gradients of a function with a finite
 #' difference approximation and prints the results of these checks.
 #'
-#' @param .x point at which the comparison is done.
-#' @param func function to be evaluated.
-#' @param func_grad function calculating the analytic gradients.
-#' @param check_derivatives_tol option determining when differences between the
-#' analytic gradient and its finite difference approximation are flagged as an
-#' error.
-#' @param check_derivatives_print option related to the amount of output. 'all'
-#' means that all comparisons are shown, 'errors' only shows comparisons that
-#' are flagged as an error, and 'none' shows the number of errors only.
-#' @param func_grad_name option to change the name of the gradient function
-#' that shows up in the output.
-#' @param ...  further arguments passed to the functions func and func_grad.
+#' @param .x Point at which the comparison is done.
+#' @param func Function to be evaluated.
+#' @param func_grad Function calculating the analytic gradients.
+#' @param check_derivatives_tol Option determining when differences between the
+#'   analytic gradient and its finite difference approximation are flagged as an
+#'   error.
+#' @param check_derivatives_print Option related to the amount of output.
+#'   \code{all} means that all comparisons are shown, \code{errors} only shows
+#'   comparisons that are flagged as an error, and \code{none} shows the number
+#'   of errors only.
+#' @param func_grad_name Option to change the name of the gradient function.
+#'   that shows up in the output.
+#' @param points Either a 3 or 5 point numerical estimate.
+#' @param \dots  Further arguments passed to the functions \code{func} or
+#'   \code{func_grad}.
 #'
 #' @return The return value contains a list with the analytic gradient, its
-#' finite difference approximation, the relative errors, and vector comparing
-#' the relative errors to the tolerance.
+#'   finite difference approximation, the relative errors, and a vector
+#'   comparing the relative errors to the tolerance.
 #'
 #' @export
 #'
-#' @author Jelmer Ypma
+#' @author Jelmer Ypma, Avraham Adler \email{Avraham.Adler@@gmail.com}
 #'
-#' @seealso \code{\link[nloptr:nloptr]{nloptr}}
+#' @seealso \code{\link[nloptr:nloptr]{nloptr}},
+#'  \code{\link[nloptr:nl.grad]{nl.grad}}
 #'
 #' @keywords optimize interface
 #'
@@ -79,6 +86,10 @@
 #' check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
 #'           check_derivatives_print = 'errors', a = runif(10))
 #'
+#' check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
+#'           check_derivatives_print = 'errors',
+#'           points = 5, a = runif(10))
+#'
 #' # example with incorrect gradient of vector-valued function
 #' g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))
 #'
@@ -90,16 +101,25 @@
 #' check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
 #'           check_derivatives_print = 'all', a = runif(10))
 #'
+
+## (2024-07-03) TO DO: Change order of parameters to put points after func_grad
+## at next breaking change (AA).
 check.derivatives <- function(.x,
                               func,
                               func_grad,
                               check_derivatives_tol = 1e-04,
                               check_derivatives_print = "all",
                               func_grad_name = "grad_f",
+                              points = 3L,
                               ...) {
+
   analytic_grad <- func_grad(.x, ...)
 
-  finite_diff_grad <- finite.diff(func, .x, ...)
+  finite_diff_grad <- if (length(func(.x, ...)) > 1L) {
+    nl.jacobian(.x, func, points = points, ...)
+  } else {
+    nl.grad(.x, func, points = points, ...)
+  }
 
   relative_error <- ifelse(finite_diff_grad == 0,
                            analytic_grad,
@@ -147,7 +167,6 @@ check.derivatives <- function(.x,
             "\n\n")
   } else if (check_derivatives_print == "errors") {
     if (sum(flag_derivative_warning) > 0) {
-
       message("\n",
               paste0(ifelse(flag_derivative_warning[flag_derivative_warning],
                             "*", " "),
@@ -163,10 +182,9 @@ check.derivatives <- function(.x,
                      "]", collapse = "\n"),
               "\n\n")
     }
-  } else if (check_derivatives_print == "none") {}
+  }
 
-  list("analytic"        = analytic_grad,
-       "finite_difference"     = finite_diff_grad,
-       "relative_error"      = relative_error,
-       "flag_derivative_warning" = flag_derivative_warning)
+  list(analytic = analytic_grad, finite_difference = finite_diff_grad,
+       relative_error = relative_error,
+       flag_derivative_warning = flag_derivative_warning)
 }